backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
boto3==1.18.63
PyPDF2==2.10.5
pdfplumber==0.7.1
google-cloud-aiplatform>=1.25.0
vertexai>=0.0.1
Werkzeug==2.0.1
pypdf==3.17.1
protobuf>=3.19.0